+++ src/client/src/pages/overview/boreholeTable/boreholeTable.tsx
@@ -23,6 +23,7 @@ import { useDomains } from "../../../api/fetchApiV2";
 import { theme } from "../../../AppTheme.ts";
 import { useAuth } from "../../../auth/useBdmsAuth.tsx";
 import { Table } from "../../../components/table/table.tsx";
+import { formatWithThousandSeparator } from "../../../utils.ts";
 import { OverViewContext } from "../overViewContext.tsx";
 
 export interface BoreholeTableProps {
@@ -37,19 +38,6 @@ export interface BoreholeTableProps {
   rowToHighlight: number | null;
   isBusy: boolean;
 }
-
-const formatWithThousandSeparator = (value: number | null): string => {
-  if (value == null) return "-";
-
-  // Format number using de-CH
-  const formatted = new Intl.NumberFormat("de-CH", {
-    useGrouping: true,
-    minimumFractionDigits: 2,
-  }).format(value);
-
-  // Ensure thousand separators are always a standard single quote (')
-  return formatted.replace(/’/g, "'");
-};
 
 export const BoreholeTable: FC<BoreholeTableProps> = ({
   boreholes,