+++ src/client/src/utils.ts
@@ -3,6 +3,19 @@ export function capitalizeFirstLetter(text: string) {
   return text.charAt(0).toUpperCase() + text.slice(1);
 }
 
+export const formatWithThousandSeparator = (value: number | null): string => {
+  if (value == null) return "-";
+
+  // Format number using de-CH
+  const formatted = new Intl.NumberFormat("de-CH", {
+    useGrouping: true,
+    minimumFractionDigits: 2,
+  }).format(value);
+
+  // Ensure thousand separators are always a standard single quote (')
+  return formatted.replace(/’/g, "'");
+};
+
 export const downloadData = (dataString: string, fileName: string, type: string) => {
   const blob = new Blob([dataString], { type: type });
   downloadDataFromBlob(blob, fileName);